- TextStyleModel 제거
