+++ Dayol-iOS/UI/Common/View/ToolBarContent/TextStyle/TextStyleAdditionalOptionView.swift
@@ -11,21 +11,36 @@ import RxCocoa
 import Combine
 
 private enum Design {
-    static let selectedButtonImage = UIImage(named: TextStyleModel.AdditionalOption.selectedImageName)
+    static let selectedButtonImage = UIImage(named: TextStyleAdditionalOptionView.AdditionalOption.selectedImageName)
     static let stackViewSpacing: CGFloat = 4.0
 }
 
 class TextStyleAdditionalOptionView: UIView {
 
+    enum AdditionalOption: String, CaseIterable {
+        case bold, cancelLine, underLine
+
+        static let selectedImageName = "toolBar_textStyle_button_selected"
+
+        var backgroundImageName: String {
+            switch self {
+            case .bold: return "toolBar_textStyle_\(rawValue)"
+            case .cancelLine: return "toolBar_textStyle_\(rawValue)"
+            case .underLine: return "toolBar_textStyle_\(rawValue)"
+            }
+        }
+
+    }
+
     private let disposeBag = DisposeBag()
-    var currentOptions: Set<TextStyleModel.AdditionalOption> = [] {
+    var currentOptions: Set<AdditionalOption> = [] {
         didSet { updateCurrentState() }
     }
-    var optionsSubject = CurrentValueSubject<Set<TextStyleModel.AdditionalOption>, Never>([])
+    var optionsSubject = CurrentValueSubject<Set<AdditionalOption>, Never>([])
     // MARK: - UI Property
 
     private let boldButton: UIButton = {
-        let image = UIImage(named: TextStyleModel.AdditionalOption.bold.backgroundImageName)
+        let image = UIImage(named: TextStyleAdditionalOptionView.AdditionalOption.bold.backgroundImageName)
         let button = UIButton()
         button.setBackgroundImage(image, for: .normal)
         button.setImage(Design.selectedButtonImage, for: .selected)
@@ -34,7 +49,7 @@ class TextStyleAdditionalOptionView: UIView {
     }()
 
     private let cancelLineButton: UIButton = {
-        let image = UIImage(named: TextStyleModel.AdditionalOption.cancelLine.backgroundImageName)
+        let image = UIImage(named: AdditionalOption.cancelLine.backgroundImageName)
         let button = UIButton()
         button.setBackgroundImage(image, for: .normal)
         button.setImage(Design.selectedButtonImage, for: .selected)
@@ -43,7 +58,7 @@ class TextStyleAdditionalOptionView: UIView {
     }()
 
     private let underLineButton: UIButton = {
-        let image = UIImage(named: TextStyleModel.AdditionalOption.underLine.backgroundImageName)
+        let image = UIImage(named: AdditionalOption.underLine.backgroundImageName)
         let button = UIButton()
         button.setBackgroundImage(image, for: .normal)
         button.setImage(Design.selectedButtonImage, for: .selected)
@@ -66,6 +81,27 @@ class TextStyleAdditionalOptionView: UIView {
 
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+
+    func setAdditionalOption(attributes: [NSAttributedString.Key: Any?]) {
+        var additionalOption: [TextStyleAdditionalOptionView.AdditionalOption] = []
+
+        if let font = attributes[.font] as? UIFont,
+           font.isBold {
+            additionalOption.append(.bold)
+        }
+
+        if let strikethroughStyle = attributes[.strikethroughStyle] as? Int,
+           strikethroughStyle == NSUnderlineStyle.single.rawValue {
+            additionalOption.append(.cancelLine)
+        }
+
+        if let underlineStyle = attributes[.underlineStyle] as? Int,
+           underlineStyle == NSUnderlineStyle.single.rawValue {
+            additionalOption.append(.underLine)
+        }
+
+        currentOptions = Set(additionalOption)
     }
 
     private func updateCurrentState() {
